Поправил типы переменных

+++ Lab1_arch/Figure.cs
@@ -13,6 +13,6 @@ namespace Lab1_arch
             return (10 * x) - Math.Log(14 * x);
         }
 
-        public abstract double Calculate(int n, double a, double b);
+        public abstract double Calculate(int n, int a, int b);
     }
 }
